Add edit_file function and use it to edit commit and PR message
This change adds an edit_file function which allows users to edit the commit and PR messages before they are
created. This will give users more control over the messages and help prevent any unintentional errors. The function
is used in the create_commit and create_pr functions to edit the respective messages before they are created.