Update PR template and add list of changes
The changes are to update the PR template and to add a numeric list of changes that were made. This will make it clear what specific changes were made in this pull request. The description of the changes includes renaming the `.env` file to `.env.example` and removing sensitive information from it, in order to maintain the security of the project and protect against potential vulnerabilities.